squashing

+++ build/lib/create_release_checksums.sh
@@ -33,17 +33,20 @@ rm -f $SHA256SUM
 rm -f $SHA512SUM
 echo "Writing artifact hashes to SHA256SUM/SHA512SUM files..."
 cd $ASSET_ROOT
-for file in $(find ${ASSET_ROOT} -type f -not -path '*\.sha[25][51][62]' -not -path '*\.docker_*' \( -path '*bin/linux*' -o -path '*bin/windows*' -o -path '*bin/darwin*' -o -name '*tar\.gz' \) ); do
+for file in $(find ${ASSET_ROOT} -type f -not -path '*\.sha[25][51][62]' -not -path '*\.docker_*' \
+     \( -path '*bin/linux*' -o -path '*bin/windows*' -o -path '*bin/darwin*' -o -name '*\.gz' -o -name '*\.ova' \) ); do
     filepath=$(realpath --relative-base=${ASSET_ROOT} $file )
     sha256sum "$filepath" | tee -a $SHA256SUM > "$file.sha256" || return 1
     sha512sum "$filepath" | tee -a $SHA512SUM > "$file.sha512" || return 1
 done
 
-mv $SHA256SUM $SHA512SUM "$ASSET_ROOT"
-sha256sum SHA256SUM > "SHA256SUM.sha256" || return 1
-sha512sum SHA256SUM > "SHA256SUM.sha512" || return 1
-sha256sum SHA512SUM > "SHA512SUM.sha256" || return 1
-sha512sum SHA512SUM > "SHA512SUM.sha512" || return 1
+if [ -f $SHA256SUM ]; then
+    mv $SHA256SUM $SHA512SUM "$ASSET_ROOT"
+    sha256sum SHA256SUM > "SHA256SUM.sha256" || return 1
+    sha512sum SHA256SUM > "SHA256SUM.sha512" || return 1
+    sha256sum SHA512SUM > "SHA512SUM.sha256" || return 1
+    sha512sum SHA512SUM > "SHA512SUM.sha512" || return 1
 
-cat SHA256SUM
+    cat SHA256SUM
+fi
 cd -
